Actuators/Notification/requirements.txt
# Python packages required for ServerMain.py
hyper
cryptography
pyjwt
paho-mqtt
requests
